Replaced Index page with working "home" page
